Fix incorrect stepCount fencepost adjustment.
stepCount needs a +1 somewhere to avoid a fencepost error. (Consider the trivial case of 5 > 5 : 872 or something else where both bounds of iterative format are the same - we still want one value output.) In its original position, it didn't match the sign of the rest of the calculation, so if we have to correct the sign of step (and stepCount), it would cause us to miss two values instead of zero. Moving the "add one" until after stepCount is nonnegative cleans this up.