Fix performance when generating very long prompt lists.
Repeated string append is quadratic because it rewrites the entire string each time. Building a list of strings and joining them all together at once is linear time because C# allocates enough RAM for the final string in one shot.

This is apparent when trying to build a list of 3 prompts x 5 scales x 10000 iterations.